+++ griptape/drivers/event_listener/griptape_cloud_event_listener_driver.py
@@ -41,3 +41,9 @@ class GriptapeCloudEventListenerDriver(BaseEventListenerDriver):
 
         response = requests.post(url=url, json=event_payload, headers=self.headers)
         response.raise_for_status()
+
+    def try_publish_event_payload_batch(self, event_payload_batch: list[dict]) -> None:
+        url = urljoin(self.base_url.strip("/"), f"/api/structure-runs/{self.structure_run_id}/events")
+
+        response = requests.post(url=url, json=event_payload_batch, headers=self.headers)
+        response.raise_for_status()